+++ ccip-cli/src/commands/manual-exec.ts
@@ -20,9 +20,7 @@
  */
 
 import {
-  type ExecutionReport,
   bigIntReplacer,
-  calculateManualExecProof,
   discoverOffRamp,
   estimateReceiveExecution,
   isSupportedTxHash,
@@ -34,9 +32,9 @@ import { type Ctx, Format } from './types.ts'
 import {
   getCtx,
   logParsedError,
-  prettyCommit,
   prettyReceipt,
   prettyRequest,
+  prettyVerifications,
   selectRequest,
   withDateTimestamp,
 } from './utils.ts'
@@ -164,36 +162,19 @@ async function manualExec(
 
   const dest = await getChain(request.lane.destChainSelector)
   const offRamp = await discoverOffRamp(source, dest, request.lane.onRamp, source)
-  const commitStore = await dest.getCommitStoreForOffRamp(offRamp)
-  const commit = await dest.getCommitReport({ ...argv, commitStore, request })
 
+  const verifications = await dest.getVerifications({ ...argv, offRamp, request })
   switch (argv.format) {
     case Format.log:
-      logger.log('commit =', commit)
+      logger.log('commit =', verifications)
       break
     case Format.pretty:
       logger.info('Commit (dest):')
-      await prettyCommit.call(ctx, dest, commit, request)
+      await prettyVerifications.call(ctx, dest, verifications, request)
       break
     case Format.json:
-      logger.info(JSON.stringify(commit, bigIntReplacer, 2))
+      logger.info(JSON.stringify(verifications, bigIntReplacer, 2))
       break
-  }
-
-  const messagesInBatch = await source.getMessagesInBatch(request, commit.report, argv)
-  const execReportProof = calculateManualExecProof(
-    messagesInBatch,
-    request.lane,
-    request.message.messageId,
-    commit.report.merkleRoot,
-    dest,
-  )
-
-  const offchainTokenData = await source.getOffchainTokenData(request)
-  const execReport: ExecutionReport = {
-    ...execReportProof,
-    message: request.message,
-    offchainTokenData,
   }
 
   if (argv.estimateGasLimit != null) {
@@ -206,7 +187,11 @@ async function manualExec(
     logger.info('Estimated gasLimit override:', estimated)
     estimated += Math.ceil((estimated * argv.estimateGasLimit) / 100)
     const origLimit = Number(
-      'gasLimit' in request.message ? request.message.gasLimit : request.message.computeUnits,
+      'ccipReceiveGasLimit' in request.message
+        ? request.message.ccipReceiveGasLimit
+        : 'gasLimit' in request.message
+          ? request.message.gasLimit
+          : request.message.computeUnits,
     )
     if (origLimit >= estimated) {
       logger.warn(
@@ -223,8 +208,10 @@ async function manualExec(
     }
   }
 
+  const input = await source.getExecutionInput({ ...argv, request, verifications })
+
   const [, wallet] = await loadChainWallet(dest, argv)
-  const receipt = await dest.executeReport({ ...argv, offRamp, execReport, wallet })
+  const receipt = await dest.execute({ ...argv, offRamp, input, wallet })
 
   switch (argv.format) {
     case Format.log: